splits the getVisualizationList into two functions to be able to test without get API requests
